Guard: implement Clone
The guard that is returned by load() generally behaves as an Arc<>; however,
unlike Arc<> it does not implement Clone on itself.  Usually it is enough
to clone the underlying Arc<>, but not if you have a trait like

trait GiveMeASmartPointer {
    type Result = Clone+Deref<Target = SomeType>;

    fn give_it_to_me(&self) -> Self::Result;
}

To support this, implement Clone on Guard itself.  For now, the trivial implementation
clones the inner Arc into a new, unprotected guard.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>