Guard: move creation of Guard from unprotected() to load()
There are two callers of unprotected() and one of them, into_inner(),
does not need the unprotected guard at all.  So return the value directly
from into_inner(), and create the Guard only when needed.

This removes an unsafe block and a ptr::read/mem::forget dance from
into_inner().

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>