Guard: move creation of Guard from unprotected() to load()
There are two callers of unprotected() and one of them, into_inner(),
does not need the unprotected guard at all.  So return the value directly
from unprotected(), which is now called inc_and_unprotect(), and create
the Guard only when needed.

This removes an unsafe block and a ptr::read/mem::forget dance from into_inner(),
which only remains separate in order to remove the #[inline] annotation from
unprotect().

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>