Avoid calling from_ptr(as_ptr())
This is fishy and not officially supported (though probably works fine).

A slight change in provided method signature of RefCnt trait is done,
but it is generally not expected for people to implement that one and
if, then not to override the method. So while technically API-breaking,
not bumping the big version for that.